// types/enums/Route.ts
export enum ROUTE {
  SPLASH = 'splash',
  LOGIN = 'login',
  REGISTER = 'register',
  FORGET_PASS = 'forget_password',
  REQUEST_OTP = 'request_otp',
  FAVORITE_GENRE = 'favorite_genre',
  TABS = 'tabs',
}

export enum TABS {
  HOME = 'home',
  MOVIE_DETAILS = 'movie_details',
  SEARCH = 'search',
  SEAT_BOOKING = 'search_booking',
  TICKET = 'tick_booking',
  USER_ACCOUNT = 'user_account',
}